requirements.txt
fastapi
uvicorn
torch
pydantic
pyjwt
